requirements.txt
# 教师档案管理系统依赖库
pandas>=1.0.0
matplotlib>=3.0.0
Pillow>=8.0.0